plugins: split by folders + test renderer
