+++ generator/src/data/plugins.types.ts
@@ -1,0 +1,11 @@
+export type Plugin = {
+  name: string;
+  url: string;
+  description: string;
+  type: PluginType;
+  framework?: PluginFramework;
+};
+
+type PluginType = 'Effect' | 'Instrument';
+
+type PluginFramework = 'JUCE' | 'Rust VST' | 'GTK';
